chore: move abi in investor-foxy to contracts package
