+++ packages/contracts/src/abis/liquidityReserve.ts
@@ -1,8 +1,12 @@
-import type { InterfaceAbi } from 'ethers'
-
-export const foxyAbi: InterfaceAbi = [
+export const LIQUIDITY_RESERVE_ABI = [
   {
-    inputs: [],
+    inputs: [
+      {
+        internalType: 'address',
+        name: '_stakingToken',
+        type: 'address',
+      },
+    ],
     stateMutability: 'nonpayable',
     type: 'constructor',
   },
@@ -35,50 +39,13 @@ export const foxyAbi: InterfaceAbi = [
     anonymous: false,
     inputs: [
       {
-        indexed: true,
-        internalType: 'uint256',
-        name: 'epoch',
-        type: 'uint256',
-      },
-      {
         indexed: false,
         internalType: 'uint256',
-        name: 'rebase',
-        type: 'uint256',
-      },
-      {
-        indexed: false,
-        internalType: 'uint256',
-        name: 'index',
+        name: 'fee',
         type: 'uint256',
       },
     ],
-    name: 'LogRebase',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: 'uint256',
-        name: 'epoch',
-        type: 'uint256',
-      },
-      {
-        indexed: false,
-        internalType: 'uint256',
-        name: 'timestamp',
-        type: 'uint256',
-      },
-      {
-        indexed: false,
-        internalType: 'uint256',
-        name: 'totalSupply',
-        type: 'uint256',
-      },
-    ],
-    name: 'LogSupply',
+    name: 'FeeChanged',
     type: 'event',
   },
   {
@@ -146,12 +113,25 @@ export const foxyAbi: InterfaceAbi = [
   },
   {
     inputs: [],
-    name: 'DOMAIN_SEPARATOR',
+    name: 'BASIS_POINTS',
     outputs: [
       {
-        internalType: 'bytes32',
+        internalType: 'uint256',
         name: '',
-        type: 'bytes32',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'MINIMUM_LIQUIDITY',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
       },
     ],
     stateMutability: 'view',
@@ -160,13 +140,26 @@ export const foxyAbi: InterfaceAbi = [
   {
     inputs: [
       {
+        internalType: 'uint256',
+        name: '_amount',
+        type: 'uint256',
+      },
+    ],
+    name: 'addLiquidity',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
         internalType: 'address',
-        name: '_owner',
+        name: 'owner',
         type: 'address',
       },
       {
         internalType: 'address',
-        name: '_spender',
+        name: 'spender',
         type: 'address',
       },
     ],
@@ -185,12 +178,12 @@ export const foxyAbi: InterfaceAbi = [
     inputs: [
       {
         internalType: 'address',
-        name: '_spender',
+        name: 'spender',
         type: 'address',
       },
       {
         internalType: 'uint256',
-        name: '_value',
+        name: 'amount',
         type: 'uint256',
       },
     ],
@@ -208,44 +201,12 @@ export const foxyAbi: InterfaceAbi = [
   {
     inputs: [
       {
-        internalType: 'uint256',
-        name: '_gons',
-        type: 'uint256',
-      },
-    ],
-    name: 'balanceForGons',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
         internalType: 'address',
-        name: '_wallet',
+        name: 'account',
         type: 'address',
       },
     ],
     name: 'balanceOf',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'circulatingSupply',
     outputs: [
       {
         internalType: 'uint256',
@@ -273,12 +234,12 @@ export const foxyAbi: InterfaceAbi = [
     inputs: [
       {
         internalType: 'address',
-        name: '_spender',
+        name: 'spender',
         type: 'address',
       },
       {
         internalType: 'uint256',
-        name: '_subtractedValue',
+        name: 'subtractedValue',
         type: 'uint256',
       },
     ],
@@ -295,7 +256,7 @@ export const foxyAbi: InterfaceAbi = [
   },
   {
     inputs: [],
-    name: 'getIndex',
+    name: 'fee',
     outputs: [
       {
         internalType: 'uint256',
@@ -322,32 +283,13 @@ export const foxyAbi: InterfaceAbi = [
   {
     inputs: [
       {
-        internalType: 'uint256',
-        name: '_amount',
-        type: 'uint256',
-      },
-    ],
-    name: 'gonsForBalance',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
         internalType: 'address',
-        name: '_spender',
+        name: 'spender',
         type: 'address',
       },
       {
         internalType: 'uint256',
-        name: '_addedValue',
+        name: 'addedValue',
         type: 'uint256',
       },
     ],
@@ -363,34 +305,20 @@ export const foxyAbi: InterfaceAbi = [
     type: 'function',
   },
   {
-    inputs: [],
-    name: 'index',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
     inputs: [
       {
         internalType: 'address',
         name: '_stakingContract',
         type: 'address',
       },
-    ],
-    name: 'initialize',
-    outputs: [
       {
-        internalType: 'bool',
-        name: '',
-        type: 'bool',
+        internalType: 'address',
+        name: '_rewardToken',
+        type: 'address',
       },
     ],
+    name: 'initialize',
+    outputs: [],
     stateMutability: 'nonpayable',
     type: 'function',
   },
@@ -408,6 +336,24 @@ export const foxyAbi: InterfaceAbi = [
     type: 'function',
   },
   {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: '_amount',
+        type: 'uint256',
+      },
+      {
+        internalType: 'address',
+        name: '_recipient',
+        type: 'address',
+      },
+    ],
+    name: 'instantUnstake',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
     inputs: [],
     name: 'name',
     outputs: [
@@ -418,68 +364,6 @@ export const foxyAbi: InterfaceAbi = [
       },
     ],
     stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'owner',
-        type: 'address',
-      },
-    ],
-    name: 'nonces',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'owner',
-        type: 'address',
-      },
-      {
-        internalType: 'address',
-        name: 'spender',
-        type: 'address',
-      },
-      {
-        internalType: 'uint256',
-        name: 'value',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'deadline',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint8',
-        name: 'v',
-        type: 'uint8',
-      },
-      {
-        internalType: 'bytes32',
-        name: 'r',
-        type: 'bytes32',
-      },
-      {
-        internalType: 'bytes32',
-        name: 's',
-        type: 'bytes32',
-      },
-    ],
-    name: 'permit',
-    outputs: [],
-    stateMutability: 'nonpayable',
     type: 'function',
   },
   {
@@ -506,73 +390,13 @@ export const foxyAbi: InterfaceAbi = [
     inputs: [
       {
         internalType: 'uint256',
-        name: '_profit',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: '_epoch',
+        name: '_amount',
         type: 'uint256',
       },
     ],
-    name: 'rebase',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
+    name: 'removeLiquidity',
+    outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    name: 'rebases',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'epoch',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'rebase',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'totalStakedBefore',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'totalStakedAfter',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'amountRebased',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'index',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'blockNumberOccurred',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
     type: 'function',
   },
   {
@@ -584,7 +408,46 @@ export const foxyAbi: InterfaceAbi = [
   },
   {
     inputs: [],
+    name: 'rewardToken',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: '_fee',
+        type: 'uint256',
+      },
+    ],
+    name: 'setFee',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [],
     name: 'stakingContract',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'stakingToken',
     outputs: [
       {
         internalType: 'address',
@@ -625,12 +488,12 @@ export const foxyAbi: InterfaceAbi = [
     inputs: [
       {
         internalType: 'address',
-        name: '_to',
+        name: 'recipient',
         type: 'address',
       },
       {
         internalType: 'uint256',
-        name: '_value',
+        name: 'amount',
         type: 'uint256',
       },
     ],
@@ -649,17 +512,17 @@ export const foxyAbi: InterfaceAbi = [
     inputs: [
       {
         internalType: 'address',
-        name: '_from',
+        name: 'sender',
         type: 'address',
       },
       {
         internalType: 'address',
-        name: '_to',
+        name: 'recipient',
         type: 'address',
       },
       {
         internalType: 'uint256',
-        name: '_value',
+        name: 'amount',
         type: 'uint256',
       },
     ],
@@ -671,6 +534,13 @@ export const foxyAbi: InterfaceAbi = [
         type: 'bool',
       },
     ],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'unstakeAllRewardTokens',
+    outputs: [],
     stateMutability: 'nonpayable',
     type: 'function',
   },
